BigDecimal check for trade

+++ business/trade/trade-api/src/main/java/study/daydayup/wolf/business/trade/api/domain/entity/contract/InstallmentTerm.java
@@ -34,8 +34,8 @@ public class InstallmentTerm implements VO {
     protected String relatedTradeNo;
 
     protected Integer period;
-    protected Integer percentage;
-    protected Integer feePercentage;
+    protected BigDecimal percentage;
+    protected BigDecimal feePercentage;
 
     protected BigDecimal amount;
     protected BigDecimal  interest;
